fix: Convert to ES Modules and set target shell-version to GNOME 45
Due to changes to the JS module system, a shell extension
can only be compatible with GNOME <= 44 OR with GNOME >= 45,
so we have to remove compatibility with older versions.

As a sidenote, extensionUtils.initTranslations()
does not have to be called explicitly anymore.
It will be called by the framework instead.

See also:
https://gjs.guide/extensions/upgrading/gnome-shell-45.html